requirements.txt
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.4
cloudevents==1.10.1
dataclasses==0.8
deprecation==2.1.0
Flask==2.0.3
functions-framework==3.5.0
google-api-core==2.8.2
google-auth==2.22.0
google-cloud-pubsub==2.13.0
googleapis-common-protos==1.56.3
grpc-google-iam-v1==0.12.4
grpcio==1.48.2
grpcio-status==1.60.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
proto-plus==1.23.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==0.20.0
requests==2.27.1
rsa==4.9
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.18
watchdog==2.3.1
Werkzeug==2.0.3
zipp==3.6.0
